增加在连接数不足的情况下，从github上下载节点文件.

+++ p2p/peer.go
@@ -299,8 +299,9 @@ func (p *Peer) readStream() {
 							continue
 						}
 					}
+
 					log.Info("readStream", "block==+======+====+=>Height", block.GetBlock().GetHeight(), "from peer", p.Addr(),
-						"block size(KB)", len(pb.Encode(block))/1024, "block hash",
+						"block size(KB)", float32(len(pb.Encode(block)))/1024, "block hash",
 						blockhash)
 					msg := (*p.nodeInfo).client.NewMessage("blockchain", pb.EventBroadcastAddBlock, &pb.BlockPid{p.GetPeerName(), block.GetBlock()})
 					err = (*p.nodeInfo).client.Send(msg, false)
